perf(moe): pass fp32 route weights to FlashInfer routed NVFP4 MoE
Restores the deferred fp32 pass-through from #689, now unconditional.
The bf16 cast was added in #716 because FlashInfer 0.6.15 rejected
anything else in UnpackedPrecomputed mode; the pinned 0.6.16 consumes
route weights at their native dtype (fp32 or bf16) with no internal
cast, and deferred callers never have the values read at all - they
apply the same fp32 tensor in the fused finalize. Drops one elementwise
cast kernel per MoE layer per step. Verified with the Inkling reference
and activation parity tests.

+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/moe/flashinfer/trtllm_nvfp4.py
@@ -313,13 +313,11 @@ if platform.is_nvidia:
         )
 
         if routed:
-            # FlashInfer's UnpackedPrecomputed mode requires bf16 route weights
-            # regardless of whether it finalizes the expert outputs. Deferred
-            # callers retain their original fp32 weights for the later fused
-            # finalize step.
+            # UnpackedPrecomputed route weights pass at their native dtype
+            # (fp32 or bf16), no cast needed since flashinfer 0.6.16.
             topk = (
                 topk_ids.to(torch.int32),
-                topk_weights.to(torch.bfloat16),
+                topk_weights,
             )
             result = trtllm_fp4_block_scale_routed_moe(
                 topk_ids=topk,
